fix gametime display: remove unnecessary zero padding
For systems that use 12 hour format, the forced zero-padding of the hour
causes undesired behavior. For example, 2:45 PM shows as 02:45 PM, which
is very confusing for someone who is used to 24 hour format. Unable to
test all configurations, but 24 and 12 hour formats seem to do what's
expected with this change (24 hour still zero-pads with 'numeric' if
defaulting).