!Unfinished! Add a feature to add the league icon
Current show_league feature replaces the background team/opponent logos
with the league logo. This version does something slightly different and
shows the league logo at the bottom center of the card in addition to
the other logos

TODO Have not tested the different configuration and static 25px spacer
is probably not the best way to do what I want to do here.